refactor: basic test
