+++ test/uglifyOptions-option.test.js
@@ -671,6 +671,31 @@ describe('when applied with uglifyOptions options', () => {
     });
   });
 
+  it('matches snapshot for `unknown` option', () => {
+    const compiler = createCompiler();
+    new UglifyJsPlugin({
+      uglifyOptions: {
+        output: {
+          unknown: true,
+        },
+      },
+    }).apply(compiler);
+
+    return compile(compiler).then((stats) => {
+      const errors = stats.compilation.errors.map(cleanErrorStack);
+      const warnings = stats.compilation.warnings.map(cleanErrorStack);
+
+      expect(errors).toMatchSnapshot('errors');
+      expect(warnings).toMatchSnapshot('warnings');
+
+      for (const file in stats.compilation.assets) {
+        if (Object.prototype.hasOwnProperty.call(stats.compilation.assets, file)) {
+          expect(stats.compilation.assets[file].source()).toMatchSnapshot(file);
+        }
+      }
+    });
+  });
+
   it('disable inline optimization by default (have a lot of problems)', () => {
     const compiler = createCompiler({
       entry: `${__dirname}/fixtures/inline-optimization.js`,
